Add files via upload
This file contains some initial functions that can be used when filling gaps for air temperature and short gaps for solar radiation. The function for solar radiation could be extended to medium and large gaps as it doesn't rely on any data being present in the day. I will continue to work on this notebook and include some testing later